requirements.txt
requests
beautifulsoup4
python-telegram-bot
mysql
mysql-connector-python